+++ bayesian_changepoint_detection/__init__.py
@@ -33,6 +33,7 @@ from .device import ensure_tensor, get_device, get_device_info, to_tensor
 from .hazard_functions import constant_hazard
 from .online_likelihoods import MultivariateT, StudentT
 from .priors import const_prior, geometric_prior, negative_binomial_prior
+from .streaming import OnlineChangepointDetector
 
 __all__ = [
     "get_device",
@@ -45,6 +46,7 @@ __all__ = [
     "get_map_changepoints",
     "compute_run_length_posterior",
     "viterbi_changepoints",
+    "OnlineChangepointDetector",
     "constant_hazard",
     "const_prior",
     "geometric_prior",